+++ CHANGES.md
@@ -9,6 +9,7 @@ Release Notes.
 * Fix NPE in handleMethodException method of apm-jdk-threadpool-plugin.
 * Support for C3P0 connection pool tracing.
 * Use a daemon thread to flush logs.
+* Fix typos in `URLParser`.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/213?closed=1)